increase graph weight capacity

+++ src/lib.rs
@@ -81,12 +81,12 @@ fn connectivity(
                     let (edge_graph, source) = multi_level_graph(i as i32, j as i32, &level_dict, scale);
                     // Convert the graph to suitable format for dijkstra
                     let graph = to_adjacency(&edge_graph);
-                    let successors = |node: &u16| -> Vec<(u16, u16)> {
+                    let successors = |node: &u16| -> Vec<(u16, u32)> {
                         graph.get(node).cloned().unwrap_or_default()
                     };
 
                     // Calculate all reachable paths; the end nodes
-                    let reachables: HashMap<u16, (u16, u16)> = dijkstra_all(&source, successors);
+                    let reachables: HashMap<u16, (u16, u32)> = dijkstra_all(&source, successors);
 
                     let mut cell_paths: Vec<(f32, f32, f32, Vec<f32>)> = Vec::with_capacity(reachables.len());
                     // let mut cell_paths: Vec<(f32, f32, f32, Vec<f32>)> = Vec::new();
